add: Added file name refactor, added level eventss
