+++ src/events/CommandEvents/handleLogsEvent.js
