+++ src/index.js
@@ -68,7 +68,7 @@ const { SpotifyPlugin } = require('@distube/spotify');
 const { SoundCloudPlugin } = require('@distube/soundcloud');
 const { YtDlpPlugin } = require('@distube/yt-dlp');
 const GiveawaysManager = require("./utils/giveaway");
-const { handleLogs } = require("./events/CommandEvents/handleLogs");
+const { handleLogs } = require("./events/CommandEvents/handleLogsEvent");
 const Logs = require('discord-logs');
 const { CaptchaGenerator } = require('captcha-canvas');
 const { createCanvas } = require('canvas');
@@ -76,8 +76,6 @@ const { checkVersion } = require('./lib/version');
 
 // Schemas //
 
-const levelSchema = require('./schemas/userLevelSystem');
-const levelschema = require('./schemas/levelSetupSystem');
 const roleSchema = require("./schemas/autoRoleSystem");
 const capschema = require('./schemas/verifySystem');
 const verifyusers = require('./schemas/verifyUsersSystem');
@@ -257,71 +255,6 @@ client.distube
         embeds: [new EmbedBuilder().setColor(client.config.embedMusic)
             .setDescription('🏁 | Queue finished!')]
     }))
-
-// Leveling System //
-
-client.on(Events.MessageCreate, async (message, err) => {
-
-    const { guild, author } = message;
-    if (message.guild === null) return;
-    const leveldata = await levelschema.findOne({ Guild: message.guild.id });
-
-    if (!leveldata || leveldata.Disabled === 'disabled') return;
-    let multiplier = 1;
-    
-    multiplier = Math.floor(leveldata.Multi);
-    
-
-    if (!guild || author.bot) return;
-
-    levelSchema.findOne({ Guild: guild.id, User: author.id}, async (err, data) => {
-
-        if (err) throw err;
-
-        if (!data) {
-            levelSchema.create({
-                Guild: guild.id,
-                User: author.id,
-                XP: 0,
-                Level: 0
-            })
-        }
-    })
-
-    const channel = message.channel;
-    const give = 1;
-    const data = await levelSchema.findOne({ Guild: guild.id, User: author.id});
-
-    if (!data) return;
-
-    const requiredXP = data.Level * data.Level * 20 + 20;
-
-    if (data.XP + give >= requiredXP) {
-
-        data.XP += give;
-        data.Level += 1;
-        await data.save();
-        
-        if (!channel) return;
-
-        const levelEmbed = new EmbedBuilder()
-        .setColor(client.config.embedLevels)
-        .setAuthor({ name: `Leveling System ${client.config.devBy}` })
-        .setTitle(`> ${client.user.username} Leveling System ${client.config.arrowEmoji}`)
-        .setDescription(`\`\`\`${author.username} has leveled up to level ${data.Level}!\`\`\``)
-        .setThumbnail(author.avatarURL({ dynamic: true }))
-        .setFooter({ text: `${author.username} Leveled Up`})
-        .setTimestamp()
-
-        await message.channel.send({ embeds: [levelEmbed] }).catch(err => client.logs.error('[LEVEL_ERROR] Error sending level up message!'));
-    } else {
-
-        if(message.member.roles.cache.find(r => r.id === leveldata.Role)) {
-            data.XP += give * multiplier;
-        } data.XP += give;
-        data.save();
-    }
-})
 
 // Auto Role System //
